requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.24.0
streamlit-drawable-canvas>=0.8.0
Pillow>=9.5.0
numpy>=1.24.0 

run> streamlit run streamlit_app.py
